Configurability; expand the parser greatly
The fragment start and end can now appear, paired, inside fragments. Previously, any pair of closing curly brackets after the opening double-curly-brackets was considered to be the end of the fragment. A closing curly bracket in the middle of a fragment prevented the fragment being matched at all.

Now, single curly brackets (or whatever you want the closing tag to be) will not cause issues, except when something in the fragment directly before the end of the fragment ends up matching the end of the fragment. For example, if fragments are marked by [[ and ]], [[[1, 2, 3]]] will match the inner two square brackets rather than the outer two. Padding the fragment ([[ [1, 2, 3] ]]) solves the problem.

Even the exact start and end tags inside fragments are fine, so long as they're paired. {{ '{{ ping }}' }} now works to demonstrate how to use your ping command.